requirements.txt
Django==1.11.15
django[argon2]
django-activeurl==0.1.12
django-appconf==1.0.2
django-classy-tags==0.8.0
django-csp==3.4
django-environ==0.4.5
django-recaptcha2==1.3.1
django-redis==4.10.0
django-elasticsearch-dsl==0.5.1
elasticsearch-dsl==6.3.0
django-widget-tweaks==1.4.3
djangorestframework==3.9.0
django-allauth==0.38.0
django-ratelimit==1.1.0
docutils==0.14
psycopg2==2.7.6
psycopg2-binary==2.7.5
raven==6.9.0
pygments-lexer-solidity==0.3.1
pygments==2.2.0
argon2-cffi==18.3.0
bleach==3.0.2
misaka==2.1.0
